+++ packages/ui-core/shared/src/lib/rich-text-editor/index.ts
@@ -2,4 +2,5 @@ export * from './rich-text-editor.component';
 export * from './rich-text-toolbar.component';
 export * from './rich-text-editor.module';
 export * from './legacy-html.util';
+export * from './legacy-html-canonical.util';
 export * from './presets';
